home and search screen file structure
