+++ navigation/index.tsx
@@ -18,8 +18,8 @@ import Colors from "../constants/Colors";
 import useColorScheme from "../hooks/useColorScheme";
 import ModalScreen from "../screens/ModalScreen";
 import NotFoundScreen from "../screens/NotFoundScreen";
-import HomeScreen from "../screens/TabOneScreen";
-import SearchScreen from "../screens/TabTwoScreen";
+import HomeScreen from "../screens/HomeScreen";
+import SearchScreen from "../screens/SearchScreen";
 import {
     RootStackParamList,
     RootTabParamList,
