+++ screens/HomeScreen.tsx
@@ -1,0 +1,31 @@
+import { StyleSheet } from "react-native";
+
+import EditScreenInfo from "../components/EditScreenInfo";
+import { Text, View } from "../components/Themed";
+import { RootTabScreenProps } from "../types";
+
+export default function HomeScreen({ navigation }: RootTabScreenProps<"Home">) {
+    return (
+        <View style={styles.container}>
+            <Text style={styles.title}>Movies</Text>
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        paddingHorizontal: 10,
+        // alignItems: "center",
+        // justifyContent: "center",
+    },
+    title: {
+        fontSize: 20,
+        fontWeight: "bold",
+    },
+    separator: {
+        marginVertical: 30,
+        height: 1,
+        width: "80%",
+    },
+});
